rename weatherCard component
