+++ src/components/weatherCard/backgroundArray.js
