+++ src/components/weatherCard/weatherCard.js
